fix(analysis): resolve Copilot review comments on circular imports
Fixes #1256

**Issue 1: Circular import in inspect_query_service.py**
- Changed `from vibe3.analysis import dag_service` to direct import
- Use `dag_service_module` to avoid circular import path
- Prevents partially-initialized module errors as package grows

**Issue 2: Bidirectional layer dependency**
- Move PR scoring models (PRDimensions, RiskLevel, RiskScore) to analysis layer
- Create `analysis/pr_scoring.py` for pure data models
- Keep scoring logic (generate_score_report, calculate_risk_score) in services layer
- analysis/__init__.py now imports models only, not generate_score_report
- services/pr_scoring_service imports models from analysis and provides logic
- Update pr_analysis_service to import from correct locations

This eliminates bidirectional dependencies between analysis and services layers
while maintaining backward compatibility.

Both issues identified by Copilot PR reviewer.
